# docs/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "api-server",
    "description" : "This is the api-server",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
    "/categories/1" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    },
    "/categories/2" : {
      "patch" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  },
                  "_id" : {
                    "type" : "integer"
                  },
                  "display_name" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n        \"name\": \"supplements\",\n        \"display_name\": \"Supplements\",\n        \"description\": \"Dietary Supplement Drugs, supplementing the diet\",\n        \"_id\": 2\n    }"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    },
    "/categories/3" : {
      "delete" : {
        "description" : "Auto generated using Swagger Inspector",
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    },
    "/categories" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "example" : "supplements"
        } ],
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "post" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  },
                  "display_name" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n        \"name\": \"delete\",\n        \"display_name\": \"delete\",\n        \"description\": \"to be deleted\"\n    }"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    }
  }
}